refactor: split base and lora trainer workers
